+++ Assets/PatchKit Patcher/Scripts/AppData/Remote/WebRequestWraps/WrapRequest.cs
@@ -1,21 +1,30 @@
 using System;
 using System.Collections;
+using System.Net;
 using System.Threading;
 using UnityEngine;
 using PatchKit.Api;
+using PatchKit.Unity.Patcher.Debug;
+using ILogger = PatchKit.Logging.ILogger;
 
 namespace PatchKit.Unity.Patcher.AppData.Remote
 {
     public class WrapRequest : IHttpWebRequest
     {
-        public const string responseEncoding = "iso-8859-2";
+        private static readonly DebugLogger DebugLogger = new DebugLogger(typeof(WrapRequest));
 
-        private EventWaitHandle _waitHandle;
+        public const string ResponseEncoding = "iso-8859-2";
 
-        public string Data { get; private set; }
-        public bool WasError { get; private set; }
+        private readonly EventWaitHandle _waitHandle;
 
-        IEnumerator JobCoroutine(string url)
+        private string _data;
+        private int _statusCode;
+
+        private bool _wasError;
+        private bool _wasTimeout;
+        private string _errorText;
+
+        private IEnumerator JobCoroutine(string url)
         {
             var www = new WWW(url);
             var start = DateTime.Now;
@@ -30,19 +39,47 @@ namespace PatchKit.Unity.Patcher.AppData.Remote
                 }
             }
 
-            if (www.isDone)
+            if (www.isDone || !string.IsNullOrEmpty(www.error))
             {
-                Data = www.text;
-            }
-            else if (!string.IsNullOrEmpty(www.error))
-            {
-                WasError = true;
-                Data = www.error;
+                _data = www.text;
+
+                try
+                {
+                    // HACK: because WWW is broken and sometimes just does not return STATUS in responseHeaders we are returning status code 200 (we can assume that status code is not an error since www.error is null or empty).
+                    if (!www.responseHeaders.ContainsKey("STATUS"))
+                    {
+                        DebugLogger.LogWarning("Response headers doesn't contain status information. Since WWW marks response as one without errors, status code is set to 200 (OK).");
+                        _statusCode = 200;
+                    }
+                    else
+                    {
+                        var status = www.responseHeaders["STATUS"];
+                        DebugLogger.Log(string.Format("Response status: {0}", status));
+                        var s = status.Split(' ');
+
+                        if (s.Length >= 3 && int.TryParse(s[1], out _statusCode))
+                        {
+                            DebugLogger.Log(string.Format("Successfully parsed status code: {0}", _statusCode));
+                        }
+                        else
+                        {
+                            // HACK: Again, we can't parse the status code (it might be in some different format) - so we simply set it to 200.
+                            DebugLogger.LogWarning(
+                                "Unable to parse status code. Since WWW marks response as one without errors, status code is set to 200 (OK).");
+                            _statusCode = 200;
+                        }
+                    }
+
+                }
+                catch (Exception e)
+                {
+                    _wasError = true;
+                    _errorText = e.ToString();
+                }
             }
             else
             {
-                WasError = true;
-                Data = "Timeout exception";
+                _wasTimeout = true;
             }
 
             yield return null;
@@ -62,12 +99,17 @@ namespace PatchKit.Unity.Patcher.AppData.Remote
         {
             _waitHandle.WaitOne();
 
-            if (WasError)
+            if (_wasTimeout)
             {
-                throw new Exception(Data);
+                throw new WebException("Timeout.", WebExceptionStatus.Timeout);
             }
 
-            return new WrapResponse(Data, responseEncoding);
+            if (_wasError)
+            {
+                throw new WebException(_errorText);
+            }
+
+            return new WrapResponse(_data, _statusCode, ResponseEncoding);
         }
     }
 }