+++ Assets/PatchKit Patcher/Scripts/AppData/Remote/WebRequestWraps/WrapResponse.cs
@@ -1,39 +1,37 @@
 using System;
 using System.IO;
-using UnityEngine;
 using PatchKit.Api;
 
 namespace PatchKit.Unity.Patcher.AppData.Remote
 {
     public class WrapResponse : IHttpWebResponse
     {
-        public WrapResponse(string data, string charset)
-        {
-            CharacterSet = charset;
-            _data = data;
-        }
+        private readonly string _data;
+        private readonly int _statusCode;
 
-        private string _data;
+        public WrapResponse(string data, int statusCode, string charset)
+        {
+            _data = data;
+            _statusCode = statusCode;
+            CharacterSet = charset;
+        }
 
         public string CharacterSet { get; private set; }
 
         public System.Net.HttpStatusCode StatusCode
         {
-            get
-            {
-                return System.Net.HttpStatusCode.OK;
-            }
+            get { return (System.Net.HttpStatusCode) _statusCode; }
         }
 
-        public System.IO.Stream GetResponseStream()
+        public Stream GetResponseStream()
         {
             if (_data == null)
             {
                 throw new NullReferenceException();
             }
 
-            MemoryStream stream = new MemoryStream();
-            StreamWriter writer = new StreamWriter(stream);
+            var stream = new MemoryStream();
+            var writer = new StreamWriter(stream);
 
             writer.Write(_data);
             writer.Flush();