+++ Assets/PatchKit Patcher/Scripts/AppUpdater/AppUpdater.cs
@@ -55,7 +55,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater
                 }
                 else
                 {
-                    DebugLogger.LogWarningFormat("Strategy caused exception, being handled by fallback: {0}, Trace: {1}", ex.Message, ex.StackTrace);
+                    DebugLogger.LogWarningFormat("Strategy caused exception, being handled by fallback: {0}, Trace: {1}", ex, ex.StackTrace);
 
                     if (!TryHandleFallback(cancellationToken))
                     {
@@ -73,6 +73,8 @@ namespace PatchKit.Unity.Patcher.AppUpdater
             {
                 return false;
             }
+
+            Context.App.ReloadTemporaryDirectories(); // FIX: Bug #724
 
             _strategy = _strategyResolver.Create(fallbackType, Context);