+++ Assets/PatchKit Patcher/Scripts/Debug/EnvironmentVariables.cs
@@ -5,5 +5,6 @@
         public const string ForceSecretEnvironmentVariable = "PK_PATCHER_FORCE_SECRET";
         public const string ForceVersionEnvironmentVariable = "PK_PATCHER_FORCE_VERSION";
         public const string KeysUrlEnvironmentVariable = "PK_PATCHER_KEYS_URL";
+        public const string MainUrlEnvironmentVariable = "PK_PATCHER_MAIN_URL";
     }
 }